modules/00_leveling/01_introduction_to_langchain/src/1-fundamentos/1-hello-world.py
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv

load_dotenv()
 
model = ChatOpenAI(model="gpt-5-nano", temperature=0.5)
message = model.invoke("hello world")

print(message.content)
